+++ code/src/cljs/sixsq/nuvla/ui/apps_component/views.cljs
@@ -346,11 +346,6 @@
          [:p "Note: ensure you have a recent installation of docker."]]))))
 
 
-(defn clear-module
-  []
-  (dispatch [::events/clear-module]))
-
-
 (defn view-edit
   []
   (let [tr            (subscribe [::i18n-subs/tr])