+++ code/src/cljs/sixsq/nuvla/ui/cimi_detail/events.cljs
@@ -4,8 +4,10 @@
             [sixsq.nuvla.ui.cimi-detail.spec :as cimi-detail-spec]
             [sixsq.nuvla.ui.cimi.events :as cimi-events]
             [sixsq.nuvla.ui.cimi.spec :as cimi-spec]
-            [sixsq.nuvla.ui.history.events :as history-events]
             [sixsq.nuvla.ui.messages.events :as messages-events]
+            [sixsq.nuvla.ui.routing.events :as routing-events]
+            [sixsq.nuvla.ui.routing.routes :as routes]
+            [sixsq.nuvla.ui.routing.utils :refer [name->href str-pathify]]
             [sixsq.nuvla.ui.utils.response :as response]))
 
 
@@ -17,8 +19,8 @@
      ::cimi-api-fx/get [resource-id #(dispatch [::set-resource %])
                         :on-error #(do
                                      (cimi-api-fx/default-get-on-error resource-id %)
-                                     (dispatch [::history-events/navigate
-                                                (str "api/" collection-name)]))]}))
+                                     (dispatch [::routing-events/navigate
+                                                (str-pathify (name->href routes/api) collection-name)]))]}))
 
 
 (reg-event-db
@@ -39,7 +41,7 @@
                                                            status (str " (" status ")"))
                                           :content message
                                           :type    :success}])
-                              (dispatch [::history-events/navigate (str "api/" collection-name)])
+                              (dispatch [::routing-events/navigate (str-pathify (name->href routes/api) collection-name)])
                               (dispatch [::cimi-events/get-results]))]}))