+++ code/src/cljs/sixsq/nuvla/ui/edges/views_clusters.cljs
@@ -1,13 +1,14 @@
 (ns sixsq.nuvla.ui.edges.views-clusters
   (:require [clojure.string :as str]
-            [re-frame.core :refer [dispatch subscribe]]
+            [re-frame.core :refer [subscribe]]
             [reagent.core :as r]
             [sixsq.nuvla.ui.edges.events :as events]
             [sixsq.nuvla.ui.edges.subs :as subs]
             [sixsq.nuvla.ui.edges.views-utils :as views-utils]
-            [sixsq.nuvla.ui.history.events :as history-events]
             [sixsq.nuvla.ui.i18n.subs :as i18n-subs]
             [sixsq.nuvla.ui.main.components :as components]
+            [sixsq.nuvla.ui.routing.routes :as routes]
+            [sixsq.nuvla.ui.routing.utils :refer [name->href]]
             [sixsq.nuvla.ui.utils.general :as general-utils]
             [sixsq.nuvla.ui.utils.semantic-ui :as ui]
             [sixsq.nuvla.ui.utils.semantic-ui-extensions :as uix]
@@ -30,13 +31,12 @@
   (let [tr (subscribe [::i18n-subs/tr])]
     (fn [{:keys [id cluster-id created managers workers nuvlabox-managers
                  nuvlabox-workers name description orchestrator status-notes] :as _nuvlabox-cluster}]
-      (let [href          (str "edges/nuvlabox-cluster/" (general-utils/id->uuid id))
+      (let [href          (name->href routes/edge-cluster-details {:uuid (general-utils/id->uuid id)})
             cluster-nodes (+ (count managers) (count workers))
             nb-per-id     (group-by :id (:resources @nuvlaboxes))
             name          (or name cluster-id)]
         [uix/Card
-         {:on-click    #(dispatch [::history-events/navigate href])
-          :href        href
+         {:href        href
           :header      [:<>
                         [ui/Icon {:className "fas fa-chart-network"}]
                         (if (> (count name) 21)