+++ code/src/cljs/sixsq/nuvla/ui/utils/values.cljs
@@ -4,8 +4,8 @@
             [clojure.string :as str]
             [markdown-to-hiccup.core :as md]
             [reagent.core :as r]
-            [sixsq.nuvla.ui.history.views :as history]
-            [sixsq.nuvla.ui.utils.semantic-ui :as ui]))
+            [sixsq.nuvla.ui.utils.semantic-ui :as ui]
+            [sixsq.nuvla.ui.utils.semantic-ui-extensions :as uix]))
 
 
 (defn href?
@@ -18,13 +18,13 @@
   "Renders a link to the API detail page associated with the href. Ignores
    other values of the map (if any)."
   [{:keys [href]}]
-  [history/link (str "api/" href) (str href)])
+  [uix/Link (str "api/" href) (str href)])
 
 
 (defn as-link
   "Renders a link to the API detail page associated with the href."
   [href & {:keys [label page]}]
-  [history/link (str (or page "api") "/" href) (or label href)])
+  [uix/Link (str (or page "api") "/" href) (or label href)])
 
 
 (defn href-coll?